+++ packages/grid/src/features/standard-table/stories/States.stories.tsx
@@ -1,4 +1,4 @@
-import { FlatButton } from "@stenajs-webui/elements";
+import { FlatButton, ResultListBannerState } from "@stenajs-webui/elements";
 import * as React from "react";
 import { StandardTable } from "../components/StandardTable";
 import {
@@ -39,3 +39,23 @@ export const _Error = () => (
     error={new Error("Could not fetch users")}
   />
 );
+
+export const BannerError = () => {
+  const bannerState: ResultListBannerState = {
+    headerText: "Default banner header error text",
+    text: "Default banner error text",
+    items: [
+      { text: "First Error" },
+      { text: "Second Error" },
+      { text: "Third Error" },
+    ],
+  };
+
+  return (
+    <StandardTable
+      items={mockedItems}
+      config={standardTableConfigForStories}
+      bannerError={bannerState}
+    />
+  );
+};