+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### StandardTable
 
+- `bannerError` is added to standard table and is prioritized before error property if given. Error is now displayed in a ResultListBanner.
+- `error and errorLabel` are now displayed in an Banner instead of ErrorScreen.
 #### Column configs now have `onKeyDown` option.
 
 ```